Extract function declarations to utils.h
