+++ .gitignore
@@ -2,3 +2,4 @@
 a.out
 client
 server
+*.o
