+++ Makefile
@@ -1,15 +1,24 @@
-.PHONY: build-server
-build-server:
-	gcc -Wall -Wextra -Wpedantic server.c -o server
+CC = gcc
+CFLAGS = -Wall -Wextra -Wpedantic -O2
 
-.PHONY: run-server
-run-server: build-server
-	./server
+.PHONY: all clean server client utils
 
-.PHONY: build-client
-build-client:
-	gcc -Wall -Wextra -Wpedantic client.c -o client
+all: server client
 
-.PHONY: run-client
-run-client:build-client
-	./client
+server: server.o utils.o
+	$(CC) $(CFLAGS) -o server server.o utils.o
+
+client: client.o utils.o
+	$(CC) $(CFLAGS) -o client client.o utils.o
+
+server.o: server.c
+	$(CC) $(CFLAGS) -c server.c
+
+client.o: client.c
+	$(CC) $(CFLAGS) -c client.c
+
+utils.o: utils.c
+	$(CC) $(CFLAGS) -c utils.c
+
+clean:
+	rm -f server client *.o
